requirements.txt
torch>=1.13.1
torchvision>=0.14.1
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.60.0
tensorboardX>=2.2
timm>=0.6.0
matplotlib>=3.3.0
scikit-learn>=1.0.0
albumentations>=1.0.0 